Keagate permission privs much improved

+++ packages/scripts/keagate.sh
@@ -3,6 +3,10 @@
 set -e
 
 # OS="$(uname -s)"
+if [ -n "$SUDO_USER" ]; then
+    HOME="$(getent passwd $SUDO_USER | cut -d: -f6)"
+fi
+
 INSTALL_DIR="$HOME"
 FOLDER_NAME="Keagate"
 REPO_LOCATION="https://github.com/dilan-dio4/$FOLDER_NAME"
@@ -16,7 +20,9 @@ NODE_ARGS=""
 #     ...do dangerous stuff
 # fi
 
-function draw_spinner() {
+# https://github.com/DevelopersToolbox/bash-spinner
+
+draw_spinner() {
     # shellcheck disable=SC1003
     local -a marks=('/' '-' '\' '|')
     local i=0
@@ -30,34 +36,15 @@ function draw_spinner() {
     done
 }
 
-# https://github.com/DevelopersToolbox/bash-spinner
-
-# -------------------------------------------------------------------------------- #
-# Start Spinner                                                                    #
-# -------------------------------------------------------------------------------- #
-# A wrapper for starting the spiner, it will pass the message if supplied, and     #
-# store the PID of the process for later termination.                              #
-# -------------------------------------------------------------------------------- #
-
-function start_spinner() {
-    message=${1:-} # Set optional message
+start_spinner() {
+    message=${1:-}              # Set optional message
     draw_spinner "${message}" & # Start the Spinner:
-    SPIN_PID=$! # Make a note of its Process ID (PID):
+    SPIN_PID=$!                 # Make a note of its Process ID (PID):
     declare -g SPIN_PID
     trap stop_spinner $(seq 0 15)
 }
 
-# -------------------------------------------------------------------------------- #
-# Draw Spinner Eval                                                                #
-# -------------------------------------------------------------------------------- #
-# Draw the actual spinner on the screen, after evaluating the command to use the   #
-# result as the message, function will sit in an infinite loop as the stop script  #
-# is used to terminate the function.                                               #
-#                                                                                  #
-# NOTE: Do not call this function directly!                                        #
-# -------------------------------------------------------------------------------- #
-
-function draw_spinner_eval() {
+draw_spinner_eval() {
     # shellcheck disable=SC1003
     local -a marks=('/' '-' '\' '|')
     local i=0
@@ -71,21 +58,14 @@ function draw_spinner_eval() {
     done
 }
 
-# -------------------------------------------------------------------------------- #
-# Start Spinner Eval                                                               #
-# -------------------------------------------------------------------------------- #
-# A wrapper for starting the spiner, it will pass the command to be evaluated, and #
-# store the PID of the process for later termination.                              #
-# -------------------------------------------------------------------------------- #
-
-function start_spinner_eval() {
+start_spinner_eval() {
     command=${1} # Set the command
     if [[ -z "${command}" ]]; then
         echo "You MUST supply a command"
         exit
     fi
     draw_spinner_eval "${command}" & # Start the Spinner:
-    SPIN_PID=$! # Make a note of its Process ID (PID):
+    SPIN_PID=$!                      # Make a note of its Process ID (PID):
     declare -g SPIN_PID
     trap stop_spinner $(seq 0 15)
 }
@@ -96,7 +76,7 @@ function start_spinner_eval() {
 # A wrapper for stopping the spinner, this simply kills off the process.           #
 # -------------------------------------------------------------------------------- #
 
-function stop_spinner() {
+stop_spinner() {
     if [[ "${SPIN_PID}" -gt 0 ]]; then
         kill -9 $SPIN_PID >/dev/null 2>&1
     fi
@@ -144,9 +124,9 @@ install_node() {
     start_spinner "Installing Node and NPM via nvm"
     curl -s -o nvm.sh https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.1/install.sh
     chmod +x ./nvm.sh
+    export NVM_DIR="$HOME/.nvm"
     ./nvm.sh >/dev/null 2>&1
     rm ./nvm.sh
-    export NVM_DIR="$HOME/.nvm"
     [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"                   # This loads nvm
     [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion" # This loads nvm bash_completion
     nvm install 16 >/dev/null 2>&1
@@ -154,13 +134,32 @@ install_node() {
     stop_spinner
 }
 
-# Install Docker - used for Mongo and Nginx
-curl -fsSL https://get.docker.com -o get-docker.sh
-start_spinner "Installing Docker"
-sudo sh get-docker.sh >/dev/null 2>&1
-rm get-docker.sh
+# https://stackoverflow.com/a/42876846
+if [[ "$EUID" = 0 ]]; then
+    keagate_debug "Privilege check: already root"
+else
+    sudo -k # make sure to ask for password on next sudo
+    if sudo true; then
+        keagate_debug "Privilege check: Correct password"
+    else
+        keagate_debug "Privilege check: wrong password"
+        echo "Wrong password. Please try again."
+        exit 1
+    fi
+fi
+
+if ! keagate_has "docker"; then
+    keagate_debug "\`Docker\` command not found. Installing..."
+    # Install Docker - used for Mongo and Nginx
+    curl -fsSL https://get.docker.com -o get-docker.sh
+    start_spinner "Installing Docker"
+    sudo sh get-docker.sh >/dev/null 2>&1
+    rm get-docker.sh
+    stop_spinner
+fi
+
+# Fix permissions issue on certain ports from `docker run`
 sudo chmod 666 /var/run/docker.sock
-stop_spinner
 
 if keagate_has "node" && keagate_has "npm"; then
     keagate_debug "Node and NPM detected. Checking versions..."
@@ -173,7 +172,7 @@ if keagate_has "node" && keagate_has "npm"; then
             install_node
         else
             keagate_echo "Please manually upgrade Node to at least version 14, then run this script again."
-            exit 0
+            exit 1
         fi
     fi
 else
@@ -187,6 +186,7 @@ if [ -d "$FOLDER_NAME" ]; then
     read -p "Folder $FOLDER_NAME/ already exists in $INSTALL_DIR. Would you like me to overwrite this? (This will preserve \`config/local.json\`) [Y/n] " -n 1 -r
     echo # (optional) move to a new line
     if [[ $REPLY =~ ^[Yy]$ ]]; then
+        keagate_debug "Caching existing local.json to a temporary file..."
         cp -f $FOLDER_NAME/config/local.json ./local.json >/dev/null 2>&1 || true
         rm -rf $FOLDER_NAME
         start_spinner "Cloning Keagate repo"
